requirements.txt
certifi>=2023.7.22
urllib3>=2.0.7
flask
flask-cors
google-generativeai
python-dotenv
requests
beautifulsoup4
